requirements.txt
accelerate==1.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
imutils==0.5.4
kiwisolver==1.4.8
matplotlib==3.10.0
MouseInfo==0.1.3
numpy==2.2.2
python-dotenv==1.0.1
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.32.3
rubicon-objc==0.5.0
scipy==1.15.1
six==1.17.0
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.20.1
transformers==4.49.0
urllib3==1.26.15
